remove most FN2 macro keybindings
These macros are all doable at the software layer, at least with rxvt-unicode.
Doing it at that layer also makes them easy to change (no need to
recompile/reload keyboard firmware to make new changes).

To port these bindings to rxvt-unicode, we need to just add entries into
~/.Xdefaults as `URxvt.keysym.*', like this:

    URxvt.keysym.Mod4-l: \ ls -ahls --color\n

this one maps the left Super key to make it type ` ls -ahls --color' followed by
<RET>. The leading space makes it so that it does not get preserved into the
shell's history (in Zsh, you can make a setting that does this with `setopt
HIST_IGNORE_SPACE').

Last but not least, the idea of having these macros was not that useful, given
the fact that (1) all macros were meant to be used only in the terminal (there
was more than 1 instance where a macro was used in a chat message, resulting in
embarrassment), and also (2) it was difficult to remember the various macros as
many of them had <RET> at the end, which meant there was no way to try out
various macros without always executing them.